6785 interpolate settings after flattening settings hash
