+++ logstash-core/lib/logstash/settings.rb
@@ -2,11 +2,14 @@
 require "logstash/util/loggable"
 require "fileutils"
 require "logstash/util/byte_value"
+require "logstash/util/environment_variables"
 require "logstash/util/time_value"
 
 module LogStash
   class Settings
 
+    include LogStash::Util::EnvironmentVariables
+    
     def initialize
       @settings = {}
       # Theses settings were loaded from the yaml file
@@ -108,8 +111,10 @@ module LogStash
 
     def from_yaml(yaml_path)
       settings = read_yaml(::File.join(yaml_path, "logstash.yml"))
-      substitute_env(settings)
-      self.merge(flatten_hash(settings), true)
+      self.merge(
+        deep_replace(flatten_hash(settings)),
+        true
+      )
       self
     end
     
@@ -148,15 +153,6 @@ module LogStash
         h.each { |k,r| flatten_hash(r,"#{f}.#{k}",g) }
       end
       g
-    end
-
-    # Substitutes environment variables into setting values of string type.
-    def substitute_env(settings)
-      settings.each do |_, setting|
-        if setting.is_a? String
-          setting.gsub!(/\$([a-zA-Z_.][a-zA-Z0-9_.]*)|\${\g<1>}|%\g<1>%/) {ENV[$1]}
-        end
-      end
     end
   end
